requirements.txt
python-telegram-bot==20.7
google-cloud-speech==2.23.0
openai==1.6.1
beautifulsoup4==4.12.2